# ab-testing-service/server/common/api.yml
openapi: 3.0.1
info:
  title: startiq-testing
  description: app for startiq a/b testing
  version: 1.0.0
servers:
- url: /api/v1
...
paths:
...
  /examples/{id}:
    get:
      tags:
      - Examples
      parameters:
      - name: id
        in: path
        description: The id of the example to retrieve
        required: true
        schema:
          type: integer
      responses:
        200:
          description: Return the example with the specified id
          content: {}
        404:
          description: Example not found
          content: {}
...